Removes old ExerciseViewModel and renames ExerciseViewModelNew to it's name
