+++ app/src/main/res/layout/exercise_fragment_layout.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.thedistantblue.triaryapp.viewmodels.ExerciseViewModelNew" />
+            type="com.thedistantblue.triaryapp.viewmodels.ExerciseViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
